requirements.txt
requests
Pillow
pytesseract